+++ test/unit/query-change-detector.test.js
@@ -27,54 +27,6 @@ if (config.platform.isNode()) {
 
 
 config.parallel('query-change-detector.test.js', () => {
-    describe('doesDocMatchQuery()', () => {
-        it('should match', async () => {
-            const col = await humansCollection.create(0);
-            const q = col.find().where('firstName').ne('foobar');
-            const docData = schemaObjects.human();
-            assert.ok(QueryChangeDetector.doesDocMatchQuery(q._queryChangeDetector, docData));
-            col.database.destroy();
-        });
-        it('should not match', async () => {
-            const col = await humansCollection.create(0);
-            const q = col.find().where('firstName').ne('foobar');
-            const docData = schemaObjects.human();
-            docData.firstName = 'foobar';
-            assert.equal(false, QueryChangeDetector.doesDocMatchQuery(q._queryChangeDetector, docData));
-            col.database.destroy();
-        });
-        it('should match ($gt)', async () => {
-            const col = await humansCollection.create(0);
-            const q = col.find().where('age').gt(1);
-            const docData = schemaObjects.human();
-            docData.age = 5;
-            assert.ok(QueryChangeDetector.doesDocMatchQuery(q._queryChangeDetector, docData));
-            col.database.destroy();
-        });
-        it('should not match ($gt)', async () => {
-            const col = await humansCollection.create(0);
-            const q = col.find().where('age').gt(100);
-            const docData = schemaObjects.human();
-            docData.age = 5;
-            assert.equal(false, QueryChangeDetector.doesDocMatchQuery(q._queryChangeDetector, docData));
-            col.database.destroy();
-        });
-        it('BUG: this should match', async () => {
-            const col = await humansCollection.create(0);
-            const q = col.find();
-
-            const docData = {
-                color: 'green',
-                hp: 100,
-                maxHP: 767,
-                name: 'asdfsadf',
-                _rev: '1-971bfd0b8749eb33b6aae7f6c0dc2cd4'
-            };
-
-            assert.equal(true, QueryChangeDetector.doesDocMatchQuery(q._queryChangeDetector, docData));
-            col.database.destroy();
-        });
-    });
     describe('._isDocInResultData()', async () => {
         it('should return true', async () => {
             const col = await humansCollection.create(5);